+++ CHANGELOG.md
@@ -9,6 +9,7 @@
   * Two options for plasma-coil forces, one that requires the full plasma flux tables and produces
     a matrix covering the full domain, and another that does a direct calculation from only the points
     inside the limiter.
+* Update gmsh settings for passive structure meshing to be more deterministic
 * Add logging
 * Add support for python 3.13